# apps/vqa-image-demo/openapi.json
{"openapi":"3.1.0","info":{"title":"Receipt CAPTCHA API","description":"영수증 캡챠 서비스 - 영수증 이미지와 Q&A 제공","version":"1.0.0"},"paths":{"/api/v1/receipts/random":{"get":{"tags":["receipts"],"summary":"Get Random Captcha","operationId":"get_random_captcha_api_v1_receipts_random_get","parameters":[{"name":"receipt_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"영수증 종류 (예: restaurant)","title":"Receipt Type"},"description":"영수증 종류 (예: restaurant)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaptchaResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/receipts":{"get":{"tags":["receipts"],"summary":"List Receipts","operationId":"list_receipts_api_v1_receipts_get","parameters":[{"name":"batch_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Batch Id"}},{"name":"receipt_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Receipt Type"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReceiptSummary"},"title":"Response List Receipts Api V1 Receipts Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/receipts/{receipt_id}":{"get":{"tags":["receipts"],"summary":"Get Receipt","operationId":"get_receipt_api_v1_receipts__receipt_id__get","parameters":[{"name":"receipt_id","in":"path","required":true,"schema":{"type":"string","title":"Receipt Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Receipt"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/receipts/{receipt_id}/questions":{"get":{"tags":["qa"],"summary":"Get Questions","operationId":"get_questions_api_v1_receipts__receipt_id__questions_get","parameters":[{"name":"receipt_id","in":"path","required":true,"schema":{"type":"string","title":"Receipt Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionOnly"},"title":"Response Get Questions Api V1 Receipts  Receipt Id  Questions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/receipts/{receipt_id}/verify":{"post":{"tags":["qa"],"summary":"Verify Answers","operationId":"verify_answers_api_v1_receipts__receipt_id__verify_post","parameters":[{"name":"receipt_id","in":"path","required":true,"schema":{"type":"string","title":"Receipt Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/internal/ai/health":{"get":{"tags":["Health"],"summary":"Ai Health Check","description":"AI health check internal endpoint","operationId":"ai_health_check_internal_ai_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/judge":{"post":{"tags":["AI Verification"],"summary":"Judge","description":"내부 봇 판별 엔드포인트 - ModuleController에서 호출됨.\nproblem_id는 qa_pairs 테이블의 question_id(int)에 해당.","operationId":"judge_judge_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JudgeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JudgeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnswerSubmit":{"properties":{"question_id":{"type":"integer","title":"Question Id"},"answer":{"type":"string","title":"Answer"}},"type":"object","required":["question_id","answer"],"title":"AnswerSubmit"},"CaptchaResponse":{"properties":{"receipt_id":{"type":"string","title":"Receipt Id"},"receipt_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Receipt Type"},"image_data":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Data"},"question_id":{"type":"integer","title":"Question Id"},"question":{"type":"string","title":"Question"}},"type":"object","required":["receipt_id","receipt_type","image_data","question_id","question"],"title":"CaptchaResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"HealthResponse"},"JudgeRequest":{"properties":{"problem_id":{"type":"string","title":"Problem Id"},"user_answer":{"type":"string","title":"User Answer"}},"type":"object","required":["problem_id","user_answer"],"title":"JudgeRequest"},"JudgeResponse":{"properties":{"is_human":{"type":"boolean","title":"Is Human"},"module":{"type":"string","title":"Module"},"passed":{"type":"boolean","title":"Passed"}},"type":"object","required":["is_human","module","passed"],"title":"JudgeResponse"},"QuestionOnly":{"properties":{"id":{"type":"integer","title":"Id"},"receipt_id":{"type":"string","title":"Receipt Id"},"question":{"type":"string","title":"Question"},"order":{"type":"integer","title":"Order","default":1}},"type":"object","required":["id","receipt_id","question"],"title":"QuestionOnly"},"Receipt":{"properties":{"id":{"type":"string","title":"Id"},"receipt_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Receipt Type","default":"restaurant"},"content":{"type":"string","title":"Content"},"image_data":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Data"},"batch_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Batch Id"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","content"],"title":"Receipt"},"ReceiptSummary":{"properties":{"id":{"type":"string","title":"Id"},"receipt_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Receipt Type","default":"restaurant"},"batch_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Batch Id"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"}},"type":"object","required":["id"],"title":"ReceiptSummary"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerifyRequest":{"properties":{"answers":{"items":{"$ref":"#/components/schemas/AnswerSubmit"},"type":"array","title":"Answers"}},"type":"object","required":["answers"],"title":"VerifyRequest"},"VerifyResponse":{"properties":{"receipt_id":{"type":"string","title":"Receipt Id"},"total":{"type":"integer","title":"Total"},"correct_count":{"type":"integer","title":"Correct Count"},"passed":{"type":"boolean","title":"Passed"},"results":{"items":{"$ref":"#/components/schemas/VerifyResult"},"type":"array","title":"Results"}},"type":"object","required":["receipt_id","total","correct_count","passed","results"],"title":"VerifyResponse"},"VerifyResult":{"properties":{"question_id":{"type":"integer","title":"Question Id"},"correct":{"type":"boolean","title":"Correct"},"expected":{"type":"string","title":"Expected"},"submitted":{"type":"string","title":"Submitted"}},"type":"object","required":["question_id","correct","expected","submitted"],"title":"VerifyResult"}}}}